test: tests for new route + moved pathfinding tests
