+++ app/test/routes/test_api_routes.py
@@ -1,229 +1,214 @@
+import datetime
 from fastapi import FastAPI
 from fastapi.testclient import TestClient
-import importlib
+import json
 import pytest
 
 # Import the router from the new structure.
 from app.routes.api_routes import router
-from app.services import pathfinding_service
-
 
 app = FastAPI()
 app.include_router(router)
 client = TestClient(app)
 
-
-@pytest.fixture(autouse=True)
-def mock_env(monkeypatch):
-	monkeypatch.setattr(
-		'app.services.pathfinding_service.os.getenv',
-		lambda key, default=None: {
-			'SENSOR_SIM': 'http://mock-sensor-sim',
-			'PATHFINDING': 'http://mock-pathfinding',
-		}.get(key, default),
-	)
-
-
-@pytest.fixture()
-def valid_room_data():
-	return {
-		'rooms': [
-			{
-				'id': 'room1',
-				'name': 'RoomA',
-				'type': 'EXHIBITION',
-				'crowd_factor': 0.5,
-				'occupants': 10,
-				'area': 100.0,
-				'longitude': 1.0,
-				'latitude': 1.0,
-				'popularity_factor': 1.0,
-			},
-			{
-				'id': 'room2',
-				'name': 'RoomB',
-				'type': 'EXHIBITION',
-				'crowd_factor': 0.5,
-				'occupants': 10,
-				'area': 100.0,
-				'longitude': 1.0,
-				'latitude': 1.0,
-				'popularity_factor': 1.0,
-			},
-		]
-	}
-
-
-@pytest.fixture
-def valid_sensor_data():
-    return {
-        "sensors": [
-            {"id": "sensor1", "rooms": ["RoomA", "RoomB"], "latitude": 55.6887823848, "longitude": 12.57792893289},
-            {"id": "sensor2", "rooms": ["RoomA", "RoomB"], "latitude": 55.6887823848, "longitude": 12.57792893289}
-        ]
-    }
-@pytest.fixture
-def valid_fastest_path_response():
-	return {"fastest_path": [
-  	{"id": "RoomA", "latitude": 55.6887823848, "longitude": 12.57792893289},
-   	{"id": "room1", "latitude": 55.6887823848, "longitude": 12.57792893289},
-   	{"id": "RoomB", "latitude": 55.6887823848, "longitude": 12.57792893289}],
-  	"distance": 10}
-
 def test_health_check():
 	response = client.get('/health')
 	assert response.status_code == 200
 	assert response.json() == {'status': 'ok'}
+ 
+ 
+def test_search_artwork_valid():
+    response = client.get("/search-artwork?keys=minimumsbetragtning")
+    assert response.status_code == 200
+    assert "Minimumsbetragtning" in response.json()
 
+def test_search_artwork_empty_string():
+    response = client.get("/search-artwork?keys=")
+    assert response.status_code == 400
+    assert response.json()["detail"] == "The 'keys' parameter must be a non-empty string."
 
-# Test: Blank source value should trigger 400 error.
-def test_blank_source():
-	payload = {'source': '   ', 'target': 'RoomA'}
-	response = client.post('/fastest-path', json=payload)
-	assert response.status_code == 422
-	assert response.json()['detail'] == [
-		{
-			'ctx': {'error': {}},
-			'input': '   ',
-			'type': 'value_error',
-			'loc': ['body', 'source'],
-			'msg': "Value error, Field 'source' must be a non-empty string.",
-		}
-	]
+def test_search_artwork_no_keys():
+    response = client.get("/search-artwork")
+    assert response.status_code == 422
 
+def test_search_artwork_invalid_json_response(monkeypatch):
+    """Simulates an invalid JSON response from the SMK API"""
 
-# Test: Blank target value should trigger 400 error.
-def test_blank_target():
-	payload = {'source': 'RoomA', 'target': '   '}
-	response = client.post('/fastest-path', json=payload)
-	assert response.status_code == 422
-	assert response.json()['detail'] == [
-		{
-			'ctx': {'error': {}},
-			'input': '   ',
-			'type': 'value_error',
-			'loc': ['body', 'target'],
-			'msg': "Value error, Field 'target' must be a non-empty string.",
-		}
-	]
+    async def mock_get(*args, **kwargs):
+        class MockResponse:
+            def raise_for_status(self):
+                pass
+            def json(self):
+                raise json.JSONDecodeError("Invalid JSON", "", 0)
 
+        return MockResponse()
 
-# Test: Simulate sensor simulation service failure (first forward_request call fails).
-def test_sensor_sim_failure(monkeypatch):
-	async def fake_forward_request(*args, **kwargs):
-		raise Exception('Sensor service error')
+    monkeypatch.setattr("httpx.AsyncClient.get", mock_get)
 
-	monkeypatch.setattr('app.services.pathfinding_service.forward_request', fake_forward_request)
+    response = client.get("/search-artwork?keys=query")
 
-	payload = {'source': 'RoomA', 'target': 'RoomB'}
-	response = client.post('/fastest-path', json=payload)
+    assert response.status_code == 500
+    assert "Internal server error" in response.json()["detail"]
+    
+    
+def test_search_artwork_api_failure(monkeypatch):
+	"""Simulates a failure from the SMK API"""
+
+	async def mock_get(*args, **kwargs):
+		class MockResponse:
+			def raise_for_status(self):
+					raise Exception("API failure")
+			def json(self):
+					return {}
+
+		return MockResponse()
+
+	monkeypatch.setattr("httpx.AsyncClient.get", mock_get)
+
+	response = client.get("/search-artwork?keys=query")
+
 	assert response.status_code == 500
-	assert 'Failed to retrieve room data' in response.json()['detail']
+	assert "Internal server error" in response.json()["detail"]
 
 
-# Test: When sensor service returns invalid room data (not a dict with 'rooms' key or a list).
-def test_invalid_room_data(monkeypatch):
-	async def fake_forward_request(*args, **kwargs):
-		return 'invalid data', 200
-
-	monkeypatch.setattr('app.services.pathfinding_service.forward_request', fake_forward_request)
-
-	payload = {'source': 'RoomA', 'target': 'RoomB'}
-	response = client.post('/fastest-path', json=payload)
-	assert response.status_code == 500
-	assert 'Invalid or empty room data' in response.json()['detail']
-
-
-# Test: Simulate failure from the pathfinding service (second forward_request call fails).
-def test_pathfinding_failure(monkeypatch, valid_room_data):
-	call_count = 0
-
-	async def fake_forward_request(url, method, body=None, params=None):
-		nonlocal call_count
-		call_count += 1
-		if call_count == 1:
-			# First call: sensor simulation returns valid room data.
-			return valid_room_data, 200
-		elif call_count == 2:
-			raise Exception('Pathfinding error')
-
-	monkeypatch.setattr('app.services.pathfinding_service.forward_request', fake_forward_request)
-
-	payload = {'source': 'RoomA', 'target': 'RoomB'}
-	response = client.post('/fastest-path', json=payload)
-	assert response.status_code == 500
-	assert (
-		'Failed to retrieve sensor data from sensor simulation service' in response.json()['detail']
-	)
-
-
-# Test: When pathfinding returns an invalid response (not a dict).
-def test_invalid_pathfinding_response(monkeypatch, valid_room_data):
-	call_count = 0
-
-	async def fake_forward_request(url, method, body=None, params=None):
-		nonlocal call_count
-		call_count += 1
-		if call_count == 1:
-			# Return valid room data.
-			return valid_room_data, 200
-		elif call_count == 2:
-			# Return an invalid response.
-			return {}, 500
-
-	monkeypatch.setattr('app.services.pathfinding_service.forward_request', fake_forward_request)
-
-	payload = {'source': 'RoomA', 'target': 'RoomB'}
-	response = client.post('/fastest-path', json=payload)
-	assert response.status_code == 500
-	assert (
-		'Invalid or empty sensor data received from sensor simulation service'
-		in response.json()['detail']
-	)
+@pytest.fixture
+def mock_api_response():
+		return {
+			"autocomplete": [
+			],
+			"found": 2,
+			"rows": 10,
+			"offset": 0,
+			"items": [
+				{
+					"current_location_name": "Room 1",
+					"dimensions": [],
+					"artist": ["Artist Name"],
+					"acquisition_date": "2023-01-01",
+					"responsible_department": "Department Name",
+					"titles": [{
+						"title": "Title 1",
+						"language": "da",
+        			}],
+					"production_date": [{
+        				"date": datetime.datetime(2023, 1, 1)
+            	}],
+					"colors": ["Red", "Blue"],
+					"techniques": ["Oil", "Acrylic"]
+				},
+				{
+					"current_location_name": "Room 2",
+					"dimensions": [],
+					"artist": ["Another Artist"],
+					"acquisition_date": "2023-02-01",
+					"responsible_department": "Another Department",
+					"titles": [{
+						"title": "Title 2",
+						"language": "en",
+		  			}],
+					"production_date": [{
+						"date": datetime.datetime(2023, 2, 1)
+					}],
+					"colors": ["Green"],
+					"techniques": ["Watercolor"]
+				}
+			]
+		}, 200
 
 
-# Test: A successful flow where both forward_request calls return valid data.
-def test_success(monkeypatch, valid_room_data, valid_sensor_data, valid_fastest_path_response):
-	call_count = 0
+@pytest.fixture
+def mock_room():
+		return {"name": "Room 1"}, 200
 
-	async def fake_forward_request(url, method, body=None, params=None):
-		nonlocal call_count
-		call_count += 1
-		if call_count == 1:
-			# Return room data as a dictionary with a "rooms" key.
-			return valid_room_data, 200
-		elif call_count == 2:
-			# return sensor data
-			return valid_sensor_data, 200
-		elif call_count == 3:
-			# Return a valid pathfinding response with the expected keys.
-			return valid_fastest_path_response, 200
-		else:
-			raise Exception(f'Unexpected API call with count {call_count}')
 
-	monkeypatch.setattr('app.services.pathfinding_service.forward_request', fake_forward_request)
+@pytest.fixture
+def mock_room_invalid():
+		return {"detail": {
+			'body': None,
+			'error': 'Invalid room id.',
+			'method': 'GET',
+			'url': 'http://localhost:8002/rooms/invalid_room_id',
+			'params': None,
+     }}, 400
 
-	payload = {'source': 'RoomA', 'target': 'RoomB'}
-	response = client.post('/fastest-path', json=payload)
+
+@pytest.fixture
+def mock_room_not_found():
+		return {"detail": {
+			'body': None,
+			'error': 'Room not found.',
+			'method': 'GET',
+			'url': 'http://localhost:8002/rooms/invalid_room_id',
+			'params': None,
+     }}, 404
+
+
+@pytest.mark.asyncio
+async def test_artwork_query(mocker, mock_api_response, mock_room):
+	payload = {
+		"keys": "Test title",
+		"room": "67efbb200b23f5290bff700f"
+	}
+
+	mocker.patch("app.services.smk_api.forward_request", side_effect=[mock_room, mock_api_response])
+ 
+	response = client.get("/artwork", params=payload)
 	assert response.status_code == 200
-	json_response = response.json()
-	assert 'fastest_path' in json_response
-	assert 'distance' in json_response
+	assert isinstance(response.json(), dict)
+ 
+	assert "items" in response.json()
+	assert len(response.json()["items"]) == 2
+ 
+@pytest.mark.asyncio
+async def test_artwork_query_no_room(mocker, mock_api_response, mock_room):
+	payload = {
+		"keys": "Test title",
+	}
+ 
+	mocker.patch("app.services.smk_api.forward_request", side_effect=[mock_api_response])
+ 
+	response = client.get("/artwork", params=payload)
+	assert response.status_code == 200
+	assert isinstance(response.json(), dict)
+ 
+	assert "items" in response.json()
+	assert len(response.json()["items"]) == 2
+ 
+ 
+@pytest.mark.asyncio
+async def test_artwork_query_invalid_room(mocker, mock_room_invalid):
+	payload = {
+		"keys": "Test title",
+		"room": "invalid_room_id"
+	}
+ 
+	mocker.patch("app.services.smk_api.forward_request", side_effect=[mock_room_invalid])
+ 
+	response = client.get("/artwork", params=payload)
+	assert response.status_code == 400
+	assert response.json()["detail"] == {
+		'body': None,
+		'error': 'Invalid room id.',
+		'method': 'GET',
+		'url': 'http://localhost:8002/rooms/invalid_room_id',
+		'params': None,
+	}
 
-
-def test_missing_pathfinding(monkeypatch):
-	monkeypatch.setattr(
-		'app.services.pathfinding_service.os.getenv',
-		lambda key, default=None: None if key == 'PATHFINDING' else default,
-	)
-	with pytest.raises(RuntimeError, match='PATHFINDING not found'):
-		importlib.reload(pathfinding_service)
-
-
-def test_missing_sensor_sim(monkeypatch):
-	monkeypatch.setattr(
-		'app.services.pathfinding_service.os.getenv',
-		lambda key, default=None: None if key == 'SENSOR_SIM' else default,
-	)
-	with pytest.raises(RuntimeError, match='SENSOR_SIM not found'):
-		importlib.reload(pathfinding_service)
+@pytest.mark.asyncio
+async def test_artwork_query_room_not_found(mocker, mock_room_not_found):
+	payload = {
+		"keys": "Test title",
+		"room": "dddddd200b23f5290bff7d0f"
+	}
+ 
+	mocker.patch("app.services.smk_api.forward_request", side_effect=[mock_room_not_found])
+ 
+	response = client.get("/artwork", params=payload)
+	assert response.status_code == 404
+	assert response.json()["detail"] == {
+		'body': None,
+		'error': 'Room not found.',
+		'method': 'GET',
+		'url': 'http://localhost:8002/rooms/invalid_room_id',
+		'params': None,
+	}
